Create Plumo.Builder

+++ plumo-core/src/main/java/org/glavo/plumo/internal/PlumoImpl.java
@@ -20,14 +20,14 @@ import org.glavo.plumo.Plumo;
 import org.glavo.plumo.internal.util.OutputWrapper;
 import org.glavo.plumo.internal.util.UnixDomainSocketUtils;
 import org.glavo.plumo.internal.util.Utils;
-import org.glavo.plumo.internal.util.VirtualThreadUtils;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLServerSocket;
 import java.io.Closeable;
 import java.io.IOException;
-import java.net.*;
-import java.nio.channels.Channels;
+import java.net.ServerSocket;
+import java.net.Socket;
+import java.net.SocketAddress;
 import java.nio.channels.ServerSocketChannel;
 import java.nio.channels.SocketChannel;
 import java.nio.file.Files;
@@ -35,134 +35,58 @@ import java.nio.file.Path;
 import java.util.Objects;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReentrantLock;
 
 public final class PlumoImpl implements Plumo {
 
-    private final ReentrantLock sessionLock = new ReentrantLock();
-    private CountDownLatch latch;
+    private final ReentrantLock lock = new ReentrantLock();
+    private final CountDownLatch latch = new CountDownLatch(1);
 
-    private static final int STATE_INIT = 0;
-    private static final int STATE_STARTUP = 1;
-    private static final int STATE_RUNNING = 2;
-    private static final int STATE_SHUTDOWN = 3;
-    private static final int STATE_COMPLETE = 4;
+    private final SocketAddress address;
+    private final Path unixDomainSocketPath;
+    private final boolean deleteUnixDomainSocketFileIfExists;
+    private final Executor executor;
+    private final boolean shutdownExecutor;
+    private final SSLContext sslContext;
+    private final String[] sslProtocols;
+    private final int timeout;
+    final HttpHandler handler;
+    private final String protocol;
 
-    private final AtomicInteger state = new AtomicInteger();
+    private volatile SocketAddress localAddress;
 
-    private SocketAddress address;
-    private Path unixDomainSocketPath;
-    private boolean deleteUnixDomainSocketFileIfExists;
-    private Executor executor;
-    private boolean shutdownExecutor;
-    private SSLContext sslContext;
-    private String[] sslProtocols;
-    private String protocol;
-    HttpHandler handler;
-    private int timeout = 5000;
+    private volatile HttpSessionImpl firstSession;
+    private volatile HttpSessionImpl lastSession;
 
-    private Closeable serverSocketOrChannel;
-    private SocketAddress localAddress;
+    private static final int STATUS_INIT = 0;
+    private static final int STATUS_RUNNING = 1;
+    private static final int STATUS_TERMINATING = 2;
+    private static final int STATUS_FINISH = 3;
 
-    private volatile HttpSessionImpl firstSession, lastSession;
+    private volatile int status = STATUS_INIT;
 
-    private void ensureNotStarted() {
-        if (state.get() != STATE_INIT) {
-            throw new IllegalStateException();
-        }
-    }
-
-    @Override
-    public void bind(InetSocketAddress address) {
-        ensureNotStarted();
-
-        Objects.requireNonNull(address);
-
-        if (this.address != null) {
-            throw new IllegalStateException("The server is already bound to address " + this.address);
-        }
-
+    public PlumoImpl(SocketAddress address, Path unixDomainSocketPath, boolean deleteUnixDomainSocketFileIfExists, Executor executor, boolean shutdownExecutor, SSLContext sslContext, String[] sslProtocols, int timeout, HttpHandler handler) {
         this.address = address;
-    }
-
-    @Override
-    public void bind(Path path, boolean deleteIfExists) {
-        ensureNotStarted();
-        UnixDomainSocketUtils.checkAvailable();
-        Objects.requireNonNull(path);
-
-        if (this.address != null) {
-            throw new IllegalStateException("The server is already bound to address " + this.address);
-        }
-
-        this.address = UnixDomainSocketUtils.createUnixDomainSocketAddress(path);
-        this.unixDomainSocketPath = path;
-        this.deleteUnixDomainSocketFileIfExists = deleteIfExists;
-    }
-
-    @Override
-    public void setSSLContext(SSLContext sslContext) {
-        ensureNotStarted();
-        Objects.requireNonNull(sslContext);
-
-        if (this.sslContext != null) {
-            throw new IllegalStateException("SSLContext has been set");
-        }
-
+        this.unixDomainSocketPath = unixDomainSocketPath;
+        this.deleteUnixDomainSocketFileIfExists = deleteUnixDomainSocketFileIfExists;
+        this.executor = executor;
+        this.shutdownExecutor = shutdownExecutor;
         this.sslContext = sslContext;
-    }
-
-    @Override
-    public void setEnabledSSLProtocols(String[] protocols) {
-        ensureNotStarted();
-        Objects.requireNonNull(protocols);
-        if (protocols.length == 0) {
-            throw new IllegalArgumentException("SSL protocols must not be empty");
-        }
-
-        if (this.sslProtocols != null) {
-            throw new IllegalStateException("SSL protocols has been set");
-        }
-
-        if (this.sslContext == null) {
-            throw new IllegalStateException("SSLContext must be set first");
-        }
-
-        this.sslProtocols = protocols;
-    }
-
-    @Override
-    public void setSocketTimeout(int timeout) {
-        ensureNotStarted();
-        if (timeout < 0) {
-            throw new IllegalArgumentException("socket timeout must not be negative");
-        }
-
+        this.sslProtocols = sslProtocols;
         this.timeout = timeout;
-    }
-
-    @Override
-    public void setHandler(HttpHandler handler) {
-        ensureNotStarted();
-        Objects.requireNonNull(handler);
-
-        if (this.handler != null) {
-            throw new IllegalStateException("Http handler has been set");
-        }
-
         this.handler = handler;
+
+        this.protocol = sslContext == null ? "http" : "https";
     }
 
-    // ---
+    private volatile Closeable serverSocketOrChannel;
 
     @Override
     public boolean isRunning() {
-        int state = this.state.get();
-        return state == STATE_RUNNING || state == STATE_SHUTDOWN;
+        int status = this.status;
+        return status > STATUS_INIT && status < STATUS_FINISH;
     }
 
     @Override
@@ -175,43 +99,70 @@ public final class PlumoImpl implements Plumo {
         return protocol;
     }
 
-    private void startImpl(ThreadFactory threadFactory) throws IOException {
-        if (!state.compareAndSet(STATE_INIT, STATE_STARTUP)) {
-            throw new IllegalStateException("Server has been started");
-        }
-
-        if (this.executor == null) {
-            final AtomicLong requestCount = new AtomicLong();
-
-            if (Constants.USE_VIRTUAL_THREAD == Boolean.TRUE || (Constants.USE_VIRTUAL_THREAD == null && VirtualThreadUtils.newThread != null)) {
-                VirtualThreadUtils.checkAvailable();
-
-                this.shutdownExecutor = false;
-                this.executor = command -> {
-                    Thread t = VirtualThreadUtils.newVirtualThread(command);
-                    t.setName("plumo-worker-" + requestCount.getAndIncrement());
-                    t.start();
-                };
-            } else {
-                this.shutdownExecutor = true;
-                this.executor = Executors.newCachedThreadPool(r -> {
-                    Thread t = new Thread(r, "plumo-worker-" + requestCount.getAndIncrement());
-                    t.setDaemon(true);
-                    return t;
-                });
-            }
-        }
-
-        if (this.handler == null) {
-            this.handler = new DefaultHttpHandler();
-        }
-
-        if (this.address == null) {
-            this.address = new InetSocketAddress(InetAddress.getLoopbackAddress(), 0);
-        }
-
+    @Override
+    public void stop() {
+        lock.lock();
         try {
-            if (this.unixDomainSocketPath == null) {
+            int status = this.status;
+
+            if (status == STATUS_INIT) {
+                this.status = STATUS_FINISH;
+            } else if (this.status == STATUS_RUNNING) {
+                this.status = STATUS_TERMINATING;
+                handler.safeClose(serverSocketOrChannel);
+            }
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    @Override
+    public void awaitTermination() throws InterruptedException {
+        if (status < STATUS_FINISH) {
+            latch.await();
+        }
+    }
+
+    @Override
+    public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
+        if (status < STATUS_FINISH) {
+            return latch.await(timeout, unit);
+        } else {
+            return true;
+        }
+    }
+
+    @Override
+    public void start(boolean daemon) throws IOException {
+        startImpl(r -> {
+            Thread thread = new Thread(r);
+            thread.setName("Plumo Listener [" + localAddress + "]");
+            if (daemon) {
+                thread.setDaemon(true);
+            }
+            return thread;
+        });
+    }
+
+    @Override
+    public void start(ThreadFactory threadFactory) throws IOException {
+        Objects.requireNonNull(threadFactory);
+        startImpl(threadFactory);
+    }
+
+    @Override
+    public void startAndWait() throws IOException {
+        startImpl(null);
+    }
+
+    private void startImpl(ThreadFactory threadFactory) throws IOException {
+        lock.lock();
+        try {
+            if (status != STATUS_INIT) {
+                throw new IllegalStateException();
+            }
+
+            if (this.unixDomainSocketPath == null && (sslContext != null || timeout > 0)) {
                 ServerSocket serverSocket;
 
                 if (sslContext == null) {
@@ -232,92 +183,50 @@ public final class PlumoImpl implements Plumo {
 
                 this.localAddress = serverSocket.getLocalSocketAddress();
             } else {
-                if (deleteUnixDomainSocketFileIfExists) {
-                    Files.deleteIfExists(unixDomainSocketPath);
+                ServerSocketChannel serverSocketChannel;
+
+                if (unixDomainSocketPath == null) {
+                    serverSocketChannel = ServerSocketChannel.open();
+                } else {
+                    if (deleteUnixDomainSocketFileIfExists) {
+                        Files.deleteIfExists(unixDomainSocketPath);
+                    }
+                    serverSocketChannel = UnixDomainSocketUtils.openUnixDomainServerSocketChannel();
                 }
 
-                ServerSocketChannel serverSocketChannel = UnixDomainSocketUtils.openUnixDomainServerSocketChannel();
                 serverSocketOrChannel = serverSocketChannel;
                 serverSocketChannel.bind(this.address);
 
                 this.localAddress = serverSocketChannel.getLocalAddress();
             }
+
+            if (threadFactory == null) {
+                this.run();
+            } else {
+                threadFactory.newThread(this::run).start();
+            }
         } catch (Throwable e) {
+            status = STATUS_FINISH;
             handler.safeClose(serverSocketOrChannel);
             serverSocketOrChannel = null;
             if (shutdownExecutor) {
                 Utils.shutdown(executor);
             }
+            lock.unlock();
             throw e;
         }
-
-        this.protocol = sslContext == null ? "http" : "https";
-        this.latch = new CountDownLatch(1);
-
-        if (threadFactory == null) {
-            this.run();
-        } else {
-            try {
-                threadFactory.newThread(this::run).start();
-            } catch (Throwable e) {
-                state.set(STATE_COMPLETE);
-                latch.countDown();
-                throw e;
-            }
-        }
     }
-
-    @Override
-    public void start() throws IOException {
-        startImpl(null);
-    }
-
-    @Override
-    public void startInNewThread(boolean daemon) throws IOException {
-        startImpl(runnable -> {
-            Thread t = new Thread(runnable);
-            t.setName("Plumo Listener [" + localAddress + "]");
-            t.setDaemon(daemon);
-            return t;
-        });
-    }
-
-    @Override
-    public void startInNewThread(ThreadFactory threadFactory) throws IOException {
-        Objects.requireNonNull(threadFactory);
-        startImpl(threadFactory);
-    }
-
-    @Override
-    public void stop() {
-        int currentState;
-        do {
-            currentState = state.get();
-            if (currentState >= STATE_SHUTDOWN) {
-                return;
-            }
-        } while (!state.compareAndSet(currentState, STATE_SHUTDOWN));
-    }
-
-    @Override
-    public void awaitTermination() throws InterruptedException {
-        if (latch != null) {
-            latch.await();
-        }
-    }
-
-    // ---
 
     private void run() {
-        if (!state.compareAndSet(STATE_STARTUP, STATE_RUNNING)) {
-            int current = state.get();
-
-            if (current == STATE_SHUTDOWN) {
-                state.compareAndSet(STATE_SHUTDOWN, STATE_COMPLETE);
-                return;
-            } else {
-                throw new AssertionError(); // ???
+        lock.lock();
+        try {
+            if (this.status != STATUS_INIT) {
+                throw new IllegalStateException("status: " + this.status);
             }
+
+            this.status = STATUS_RUNNING;
+        } finally {
+            lock.unlock();
         }
 
         try {
@@ -329,10 +238,12 @@ public final class PlumoImpl implements Plumo {
                         if (timeout > 0) {
                             socket.setSoTimeout(timeout);
                         }
-                        exec(new HttpSessionImpl(this, socket,
+                        if (!exec(new HttpSessionImpl(this, socket,
                                 socket.getRemoteSocketAddress(), socket.getLocalSocketAddress(),
                                 new HttpRequestReader(socket.getInputStream()),
-                                new OutputWrapper(socket.getOutputStream(), 1024)));
+                                new OutputWrapper(socket.getOutputStream(), 1024)))) {
+                            break;
+                        }
                     } catch (IOException e) {
                         DefaultLogger.log(DefaultLogger.Level.INFO, "Communication with the client broken", e);
                     }
@@ -343,10 +254,12 @@ public final class PlumoImpl implements Plumo {
                 do {
                     try {
                         final SocketChannel socketChannel = serverSocketChannel.accept();
-                        exec(new HttpSessionImpl(this, socketChannel,
+                        if (!exec(new HttpSessionImpl(this, socketChannel,
                                 socketChannel.getRemoteAddress(), socketChannel.getLocalAddress(),
                                 new HttpRequestReader(socketChannel),
-                                new OutputWrapper(socketChannel, 1024)));
+                                new OutputWrapper(socketChannel, 1024)))) {
+                            break;
+                        }
                     } catch (IOException e) {
                         DefaultLogger.log(DefaultLogger.Level.INFO, "Communication with the client broken", e);
                     }
@@ -358,15 +271,12 @@ public final class PlumoImpl implements Plumo {
     }
 
     private void finish() {
-        sessionLock.lock();
+        lock.lock();
         try {
-            int currentState;
-            do {
-                currentState = state.get();
-                if (currentState == STATE_COMPLETE) {
-                    return;
-                }
-            } while (!state.compareAndSet(currentState, STATE_COMPLETE));
+            if (status == STATUS_FINISH) {
+                return;
+            }
+            status = STATUS_FINISH;
 
             HttpSessionImpl session = firstSession;
 
@@ -391,23 +301,16 @@ public final class PlumoImpl implements Plumo {
                 }
             }
         } finally {
+            lock.unlock();
             latch.countDown();
-            sessionLock.unlock();
         }
     }
 
-    void exec(HttpSessionImpl session) throws IOException {
-        sessionLock.lock();
+    private boolean exec(HttpSessionImpl session) throws IOException {
+        lock.lock();
         try {
-            int currentState = state.get();
-            switch (currentState) {
-                case STATE_RUNNING:
-                    break;
-                case STATE_SHUTDOWN:
-                    // break the main loop
-                    throw new IOException("Server stopped");
-                default:
-                    throw new AssertionError("impossible state: " + currentState);
+            if (status != STATUS_RUNNING) {
+                return false;
             }
 
             HttpSessionImpl last = this.lastSession;
@@ -421,19 +324,16 @@ public final class PlumoImpl implements Plumo {
 
             lastSession = session;
         } finally {
-            sessionLock.unlock();
+            lock.unlock();
         }
 
         executor.execute(session);
+        return true;
     }
 
     void close(HttpSessionImpl session) {
-        sessionLock.lock();
+        lock.lock();
         try {
-            if (state.get() == STATE_COMPLETE) {
-                return;
-            }
-
             HttpSessionImpl next = session.next;
             HttpSessionImpl prev = session.prev;
 
@@ -451,7 +351,7 @@ public final class PlumoImpl implements Plumo {
                 session.next = null;
             }
         } finally {
-            sessionLock.unlock();
+            lock.unlock();
         }
 
         session.close();
